Restructure views for component sharing
