+++ project/package.json
@@ -44,6 +44,7 @@
     "react-router-dom": "4.3.1",
     "react-test-renderer": "^16.6.0",
     "sass-loader": "^7.1.0",
+    "styled-components": "^4.0.3",
     "webpack": "^4.24.0",
     "webpack-cli": "3.1.2",
     "webpack-dev-server": "3.1.10"
