+++ project/src/views/Restaurants/Detail/main.scss
@@ -1,0 +1,4 @@
+.leaflet-container {
+  width: 600px;
+  height: 250px;
+}
