+++ project/src/views/Restaurants/components/OpenClosed/index.js
@@ -1,0 +1,18 @@
+import React from "react";
+require("./main.scss");
+
+export const OpenNow = () => {
+  return (
+    <span>
+      <span className="green-dot" /> OPEN NOW
+    </span>
+  );
+};
+
+export const Closed = () => {
+  return (
+    <span>
+      <span className="red-dot" /> CLOSED
+    </span>
+  );
+};
